+++ src/main/java/org/folio/ld/dictionary/label/LabelGenerator.java
@@ -1,0 +1,9 @@
+package org.folio.ld.dictionary.label;
+
+import org.folio.ld.dictionary.model.Resource;
+
+public interface LabelGenerator {
+  boolean supports(Resource resource);
+
+  String getLabel(Resource resource);
+}