+++ src/main/java/org/folio/ld/dictionary/label/generators/ConceptLabelGenerator.java
@@ -1,4 +1,4 @@
-package org.folio.ld.dictionary.util;
+package org.folio.ld.dictionary.label.generators;
 
 import static java.util.Comparator.comparing;
 import static org.folio.ld.dictionary.PredicateDictionary.FOCUS;
@@ -6,33 +6,22 @@ import static org.folio.ld.dictionary.PredicateDictionary.SUB_FOCUS;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.CONCEPT;
 
 import java.util.Comparator;
-import java.util.Map;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import lombok.experimental.UtilityClass;
-import org.folio.ld.dictionary.ResourceTypeDictionary;
+import org.folio.ld.dictionary.label.LabelGenerator;
 import org.folio.ld.dictionary.model.Resource;
 import org.folio.ld.dictionary.model.ResourceEdge;
 
-@UtilityClass
-public class LabelGenerator {
+public class ConceptLabelGenerator implements LabelGenerator {
   private static final String SEPARATOR_CONCEPT = " -- ";
 
-  private static final Map<ResourceTypeDictionary, Function<Resource, String>> LABEL_GENERATORS_PER_TYPE = Map.of(
-    CONCEPT, conceptLabelGenerator()
-  );
-
-  public static String generateLabel(Resource resource) {
-    return resource.getTypes()
-      .stream()
-      .filter(LABEL_GENERATORS_PER_TYPE::containsKey)
-      .findFirst()
-      .map(t -> LABEL_GENERATORS_PER_TYPE.get(t).apply(resource))
-      .orElseGet(resource::getLabel);
+  @Override
+  public boolean supports(Resource resource) {
+    return resource.isOfType(CONCEPT);
   }
 
-  private static Function<Resource, String> conceptLabelGenerator() {
-    return resource -> resource.getOutgoingEdges()
+  @Override
+  public String getLabel(Resource resource) {
+    return resource.getOutgoingEdges()
       .stream()
       .filter(re -> re.getPredicate() == FOCUS || re.getPredicate() == SUB_FOCUS)
       .sorted((o1, o2) -> {
@@ -44,5 +33,4 @@ public class LabelGenerator {
       .map(Resource::getLabel)
       .collect(Collectors.joining(SEPARATOR_CONCEPT));
   }
-
 }