+++ NEWS.md
@@ -27,6 +27,7 @@
 - LabelGenerator and MOCKED_RESOURCE type are introduced [MODLD-789](https://folio-org.atlassian.net/browse/MODLD-789)
 - Update the IRI of ID_LOCAL to http://library.link/identifier/LOCAL [MODLD-939](https://folio-org.atlassian.net/browse/MODLD-939)
 - Add resource types FAST, GND, LCCSH, LCDGT, LCGFT, LCMPT, LCNAF, LCSH, MESH, VIAF and WIKIID [MODLD-939](https://folio-org.atlassian.net/browse/MODLD-939)
+- Create HubLabelGenerator to generate labels for Hubs [MODLD-970](https://folio-org.atlassian.net/browse/MODLD-970)
 
 ## 1.0.1 (03-12-2025)
 - Initial release